requirements-light.txt
# 轻量级OCR方案（用于本地测试）
# 使用Tesseract OCR代替PaddleOCR

# FastAPI和Web服务器
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OCR引擎（Tesseract）
pytesseract>=0.3.10
Pillow>=10.0.0

# HTTP请求
requests>=2.31.0

# 数据验证
pydantic>=2.5.0

# 日志和工具
python-dotenv>=1.0.0
